requirements.txt
# Core dependencies
pandas>=2.2.0  # Updated for Python 3.13 compatibility
numpy>=1.26.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0  # For reading Excel .xlsm files

# SQL Database connectors (install as needed)
pymysql>=1.0.0  # For MySQL
psycopg2-binary>=2.9.0  # For PostgreSQL
# pyodbc>=4.0.0  # For SQL Server (uncomment if needed)

# GenAI and LLM
openai>=1.0.0
langchain>=0.0.300
langchain-openai>=0.0.2
transformers>=4.30.0
torch>=2.0.0

# Web interface
streamlit==1.29.0

# Data visualization and analysis
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Database (sqlite3 is built-in to Python)

# Evaluation (optional - scikit-learn has Python 3.13 compatibility issues)
# scikit-learn>=1.3.0  # Uncomment if needed and using Python < 3.13
nltk>=3.8.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
tqdm==4.66.1

